Add application icon to project
The changes in the `PngToIcoConverter.csproj` file include the addition of an application icon and the inclusion of the icon file in the project. Specifically, the `<ApplicationIcon>` element is added to the `<PropertyGroup>` to specify `PNGtoICO.ico` as the application icon. Additionally, a new `<ItemGroup>` is added to include the `PNGtoICO.ico` file as content in the project.